requirements.txt
fastapi
uvicorn
pandas
scikit-surprise
boto3
